feat: Add in link back to article collection in series TOC
Fixes #7
